+++ stateMachine/src/tests.cpp
@@ -12,7 +12,7 @@ void testSwap() {
     int32_t blah2 = blah;
     swapUint(blah);
     swapUint((uint32_t &) blah2);
-    assert(blah == blah2);
+    assert(blah == (uint32_t) blah2);
     assert(blah == 0xabcdbeef);
     debugPrintf("Blah %u %d %u\n", blah, blah2, blah2);
 }
@@ -40,25 +40,25 @@ void testInteger() {
     volatile uint32_t test = (1 << 31);
     int32_t testInt = test;
     (void) testInt;
-    assert(test == testInt);
+    assert(test == (uint32_t) testInt);
     debugPrintf("Unsigned %u, signed %d\n", test, testInt);
 
     test = (1 << 31) + (1 << 30);
     testInt = test;
     (void) testInt;
-    assert(test == testInt);
+    assert(test == (uint32_t) testInt);
     debugPrintf("Unsigned %u, signed %d\n", test, testInt);
 
     test = (1 << 30);
     testInt = test;
     (void) testInt;
-    assert(test == testInt);
+    assert(test == (uint32_t) testInt);
     debugPrintf("Unsigned %u, signed %d\n", test, testInt);
 
     volatile uint64_t test64 = (1 << 31);
     int64_t testInt64 = test64;
     (void) testInt64;
-    assert(test64 == testInt64);
+    assert(test64 == (uint64_t) testInt64);
     debugPrintf("Unsigned %u, signed %d\n", test64, testInt64);
 }
